+++ application/views/_buyButton.php
@@ -1,0 +1,13 @@
+<?php
+defined('BASEPATH') OR exit('No direct script access allowed');
+/**
+ * Welcome message for logged in users
+ */
+?>
+<div id="buyCards">
+	<form method="post">
+		Purchase Card Pack?<br />
+		Cost:  <b>20 Peanuts</b><br />
+		<input type="submit" name="buyCardPack" id="buyButton" value="Purchase Card Pack"/>
+	</form>
+</div>