+++ botcards_setup.sql
@@ -33,8 +33,9 @@ CREATE TABLE IF NOT EXISTS `agent` (
 	`auth_token`		VARCHAR(100)								NOT NULL,
 	`code`				VARCHAR(5)									NOT NULL,
 	`name`				VARCHAR(100)								NOT NULL,
-	`last_updated`		DATETIME		DEFAULT CURRENT_TIMESTAMP	NOT NULL,
+	`date_registered`	DATETIME		DEFAULT CURRENT_TIMESTAMP	NOT NULL,
 	`round_registered`	INT(10)										NOT NULL,
+	`last_active_round`	INT(10)										NOT NULL,
 	CONSTRAINT `agent_pk`
 		PRIMARY KEY(`auth_token`)
 );